+++ packages/agtree/src/index.ts
@@ -7,6 +7,7 @@ export { RuleParser } from './parser/rule-parser';
 export { RuleSerializer } from './serializer/rule-serializer';
 export { RuleDeserializer } from './deserializer/rule-deserializer';
 export { RuleGenerator } from './generator/rule-generator';
+export { type OnParseError, defaultParserOptions, type ParserOptions } from './parser/options';
 
 export {
     type Agent,
@@ -111,7 +112,7 @@ export {
     UBO_SCRIPTLET_MASK,
 } from './utils/constants';
 export { BINARY_SCHEMA_VERSION } from './utils/binary-schema-version';
-export { AdblockSyntax } from './utils/adblockers';
+export { AdblockSyntax, AdblockProduct, getHumanReadableProductName } from './utils/adblockers';
 export { type CosmeticRuleSeparatorFinderResult, CosmeticRuleSeparatorUtils } from './utils/cosmetic-rule-separator';
 export { DomainUtils } from './utils/domain';
 export { type VariableTable, LogicalExpressionUtils } from './utils/logical-expression';
@@ -138,7 +139,13 @@ export { decodeTextPolyfill } from './utils/text-decoder-polyfill';
 export { RuleCategorizer } from './utils/categorizer';
 
 // Constants
-export { EXT_CSS_PSEUDO_CLASSES, EXT_CSS_LEGACY_ATTRIBUTES, FORBIDDEN_CSS_FUNCTIONS } from './converter/data/css';
+export {
+    NATIVE_CSS_PSEUDO_CLASSES,
+    REMOVE_PROPERTY,
+} from './converter/data/css';
+
+// CSS utilities
+export { CssTokenStream } from './parser/css/css-token-stream';
 
 export {
     GenericPlatform,
@@ -155,9 +162,21 @@ export {
     isGenericPlatform,
     getPlatformId,
     getSpecificPlatformName,
+    hasPlatformMultipleProducts,
+    getProductGenericPlatforms,
+    getProductSpecificPlatforms,
+    platformToAdblockProduct,
+    getPlatformsByProduct,
+    getHumanReadablePlatformName,
+    getAllPlatformNames,
     ResourceType,
     getResourceTypeModifier,
     isValidResourceType,
+    stringifyPlatforms,
+    type PlatformsByProduct,
+    type AnyPlatform,
+    PLATFORM_NEGATION,
+    PLATFORM_SEPARATOR,
 } from './compatibility-tables';
 
 // Version