+++ packages/adguard-api/CHANGELOG.md
@@ -5,11 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [3.1.1] - 2025-12-01
 
 ### Changed
 
 - Updated [@adguard/assistant] to `v4.3.77`.
+- Updated [@adguard/tswebextension] to `v3.2.17`.
+- Updated [@adguard/filters-downloader] to `v2.4.2`.
+
+[3.1.1]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/adguard-api-v3.1.1
 
 ## [3.1.0] - 2025-05-27