+++ packages/tsurlfilter/CHANGELOG.md
@@ -10,14 +10,64 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support of `$header` modifier in declarative converter.
+- Support of `:has()` pseudo-class as standard CSS conditionally:
+    - if there are no other extended pseudo-classes in the selector, or
+    - if the `#?#` separator is explicitly used in the rule [AdguardBrowserExtension#2587].
+
+### Changed
+
+<!-- TODO: update the version later -->
+- Updated [@adguard/agtree] to `vX.X.X`.
 
 ### Fixed
 
 - Apply $document blocking rule even if specific exception rule is present [AdguardBrowserExtension#3262].
-- The extension got an unexpected error on [::]:8000 [AdguardBrowserExtension#3360].
+- The extension got an unexpected error on `[::]:8000` [AdguardBrowserExtension#3360].
 
+[AdguardBrowserExtension#2587]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2587
 [AdguardBrowserExtension#3262]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3262
 [AdguardBrowserExtension#3360]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3360
+
+## [3.5.1] - 2025-11-26
+
+### Changed
+
+- Updated [@adguard/agtree] to `v3.3.1`.
+- Updated [@adguard/scriptlets] to `v2.2.13`.
+
+[3.5.1]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tsurlfilter-v3.5.1
+
+## [3.5.0] - 2025-11-20
+
+### Added
+
+- Support for regexp patterns in cosmetic rules.
+- Support for regexp in `domain` modifier [#105].
+
+### Changed
+
+- Updated [@adguard/agtree] to `v3.3.0`.
+
+[3.5.0]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tsurlfilter-v3.5.0
+[#105]: https://github.com/AdguardTeam/tsurlfilter/issues/105
+
+## [3.4.8] - 2025-11-12
+
+### Changed
+
+- Updated [@adguard/agtree] to `v3.2.5`.
+- Updated [@adguard/scriptlets] to `v2.2.12`.
+
+[3.4.8]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tsurlfilter-v3.4.8
+
+## [3.4.7] - 2025-10-17
+
+### Changed
+
+- Updated [@adguard/agtree] to `v3.2.4`.
+- Updated [@adguard/scriptlets] to `v2.2.11`.
+
+[3.4.7]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tsurlfilter-v3.4.7
 
 ## [3.4.6] - 2025-09-11